The rest of the owl

+++ SwinLocks/ecs/Components/HealthComponent.cs
@@ -6,9 +6,9 @@ using System.Threading.Tasks;
 
 namespace SwinLocks {
     class HealthComponent : Component {
-        public int health;
+        public float health;
 
-        public HealthComponent(int health)
+        public HealthComponent(float health)
             => this.health = health;
     }
 }
